requirements.txt
selenium==4.15.2
pytest==7.4.3
pytest-html==4.1.1
pytest-cov==4.1.0
allure-pytest==2.13.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
requests==2.31.0
